refactor: 팀원 및 PM 추가 다이얼로그 컴포넌트 분리 (CM-92)
